ProdNameInputBox와 ProdDescriptionInputBox 컴포넌트화 마침
