+++ sprint6/src/pages/RegisterItemPage/components/ProdDescriptionInputBox/index.jsx
@@ -1,0 +1,11 @@
+
+const ProdDescriptionInputBox = () => {
+  return (
+    <div id="prodIntroBox" className="mainWidth">
+    <h1>상품 소개</h1>
+    <input className='inputBox' type="text" placeholder="상품 소개를 입력해주세요" />
+  </div>
+  )
+}
+
+export default ProdDescriptionInputBox;
