+++ sprint6/src/pages/RegisterItemPage/index.jsx
@@ -1,16 +1,20 @@
 import { useState }  from "react"
 import "./index.css";
 import "../../styles/global.css";
+// jsx
 import Header from "./components/Header/index.jsx";
 import Footer from "../../components/Footer/index.jsx";
+import ProdNameInputBox from "./components/ProdNameInputBox/index.jsx";
+import ProdDescriptionInputBox from "./components/ProdDescriptionInputBox/index.jsx";
+// img
 import Ximg from "../../img/icons/ic_X.png";
 
 function RegisterItemPage() {
-  const [productTitleInputValue, setProductTitleInputValue] = useState();
-  const [priceInputValue, setPriceInputValue] = useState();
+  const [nameInputValue, setNameInputValue] = useState();
+  const [descriptionInputValue, setDescriptionInputValue] = useState();
   
-  const handleProductTitleOnChange = (e) => setProductTitleInputValue(e.target.value)
-  const handleProductPriceValue = (e) => setPriceInputValue(e.target.value)
+  const handleProductNameOnChange = (e) => setNameInputValue(e.target.value)
+  const handleProductDescriptionValue = (e) => setDescriptionInputValue(e.target.value)
 
   /*
   api를 쏴서, 현재 있는 input들을 백엔드쪽으로 보내는 것.
@@ -27,11 +31,14 @@ function RegisterItemPage() {
           </div>
 
           <ProdNameInputBox
-            inputValue={productTitleInputValue}
-            onChange={handleProductTitleOnChange}
+            inputValue={nameInputValue}
+            onChange={handleProductNameOnChange}
           />
 
-          <ProdDescriptionInputBox />
+          <ProdDescriptionInputBox
+            inputValue={descriptionInputValue}
+            onChange={handleProductDescriptionValue}
+          />
 
           <div id="prodPriceBox" className="mainWidth">
             <h1>판매가격</h1>
@@ -57,31 +64,6 @@ function RegisterItemPage() {
       <Footer />
     </>
   );
-}
-
-const ProdDescriptionInputBox = () => {
-  return (
-    <div id="prodIntroBox" className="mainWidth">
-    <h1>상품 소개</h1>
-    <input className='inputBox' type="text" placeholder="상품 소개를 입력해주세요" />
-  </div>
-  )
-}
-
-
-const ProdNameInputBox = ({
-  inputValue,
-  onChange
-}) => {
-  return (
-    <div id="prodNmaeBox" className="mainWidth">
-    <h1>상품명</h1>
-    <input
-    value={inputValue}
-    onChange={onChange}
-    className='inputBox' type="text" placeholder="상품명을 입력해주세요" />
-  </div>
-  )
 }
 
 export default RegisterItemPage;
